+++ pages/_app.js
@@ -145,7 +145,7 @@ function MyApp({ Component, pageProps }) {
                     placeholder="Search datasets"
                     className="search-textfield"
                   />
-                  <button type="submit" className="search-submit">
+                  <button className="search-submit">
                     <svg
                       xmlns="http://www.w3.org/2000/svg"
                       enableBackground="new 0 0 17 17"
