+++ pages/_document.js
@@ -1,4 +1,4 @@
-import { Html, Head, Main, NextScript } from 'next/document'
+import { Html, Head, Main, NextScript } from 'next/document';
 
 export default function Document() {
   return (
@@ -7,9 +7,7 @@ export default function Document() {
       <body>
         <Main />
         <NextScript />
-      <script type="text/javascript" async="" src="../scripts/custom.js"></script>
-      <script type="text/javascript" async="" src="../scripts/page-navigation.js"></script>
-      <script type="text/javascript" async="" src="../scripts/site-navigation.js"></script>
+      <script type="text/javascript" async="" src="../scripts/script.js"></script>
       </body>
     </Html>
   )
