+++ pages/preview.js
@@ -1,5 +1,5 @@
-import React from 'react';
 import { DataGrid } from '@mui/x-data-grid';
+import React, { useEffect } from 'react';
 
 export async function getServerSideProps(context) {
   function ValidateSize(FileSize) {
@@ -101,8 +101,42 @@ export async function getServerSideProps(context) {
   };
 }
 
-export default function preview(dataset) {
-  console.log()
+export default function Preview(dataset) {
+  useEffect(() => {
+    //**-- api modal start
+    var modal = document.getElementById("myModal");
+    var span = document.getElementsByClassName("close")[0];
+    span.onclick = function() {
+      modal.style.display = "none";
+    }
+
+    //add file id to api query inputs
+    function addFileId(fileId) {
+      document.getElementById("simple-query").value = `https://data.ca.gov/api/3/action/datastore_search?resource_id=${fileId}&limit=5`
+      document.getElementById("sql-query").value = `https://data.ca.gov/api/3/action/datastore_search_sql?sql=select * from "${fileId}" LIMIT 5`
+      document.getElementById("odata-query").value = `https://data.ca.gov/datastore/odata3.0/${fileId}`
+    }
+
+    const apiButtons = document.querySelectorAll('.api-button');
+    apiButtons.forEach(el => el.addEventListener('click', event => {
+      var file_id = event.target.dataset.fileId;
+      var resource_name = event.target.dataset.resourceName;
+      addFileId(file_id)
+      document.getElementById("resource-name").innerHTML = resource_name
+      modal.style.display = "block";
+    }));
+
+    //copy inputs to clipboard
+    document.querySelectorAll('.copy-button').forEach(el => el.addEventListener('click', event => {
+      event.target.classList.add("copied");
+      event.target.parentNode.querySelectorAll('input')[0].select();
+      document.execCommand("copy");
+      setTimeout(() => {
+        event.target.classList.remove("copied");
+      }, "750")
+    }));
+    //api modal end --**
+  }, []);
   return (
     <main id="body-content" className="cagov-main dataset-preview">
       <nav className="nav-breadcrumb">
