+++ pages/topic/datasets.js
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/router';
 import BasicSelect from '../../components/BasicSelect';
+import Link from 'next/link';
+
 export async function getServerSideProps(context) {
   return getFormattedData(context);
 }
@@ -735,9 +737,11 @@ const Results = (data) => {
                   className="result"
                 >
                   <h2 style={{ marginBottom: '5px' }} className="h5">
-                    <a href={"/dataset?name=" + dataset.name}>
-                      <span style={{ fontWeight: '700', fontSize: '18px', lineHeight: '32px', color: '#046A99' }}>{dataset.title}</span>
-                    </a>
+                    <Link href={"/dataset?name=" + dataset.name} passHref>
+                      <a>
+                        <span style={{ fontWeight: '700', fontSize: '18px', lineHeight: '32px', color: '#046A99' }}>{dataset.title}</span>
+                      </a>
+                    </Link>
                   </h2>
                   <ul className="result-dataset-info">
                     <li>
