+++ public/scripts/script.js
@@ -1,3 +1,291 @@
+window.addEventListener('load', function (event) {
+  if (document.getElementsByTagName('main').length > 0) {
+    if (document.getElementsByTagName('main')[0].classList.contains('home-page')) {
+      document.querySelectorAll('.site-header')[0].style.border = "0px"
+    }
+  }
+  if (document.getElementsByTagName('main').length > 0) {
+    if (document.getElementsByTagName('main')[0].classList.contains('dataset') || document.getElementsByTagName('main')[0].classList.contains('dataset-preview')) {
+      
+      //remove read more button, if no data
+      if (document.getElementById("dataset-description").getElementsByTagName('p')[0].offsetHeight === document.getElementById("dataset-description").getElementsByTagName('p')[0].scrollHeight) {
+        document.getElementsByClassName('btn-read-more')[0].style.display = "none"
+      }
+      if (document.getElementById("dataset-description").getElementsByTagName('p')[0].offsetHeight == 28) {
+        document.getElementsByClassName('btn-read-more')[0].style.display = "none"
+      }
+
+      document.querySelectorAll('.resource-description').forEach(element => {
+        if (element.parentNode.querySelectorAll('p')[0].innerHTML == "" || element.parentNode.querySelectorAll('p')[0].innerHTML == "N/A" || element.parentNode.querySelectorAll('p')[0].offsetHeight == 28) {
+          element.parentNode.querySelectorAll('.btn-read-more')[0].style.display = "none"
+        };
+      });
+
+      
+    }
+  }
+})
+
+/* remove yellow border from homepage site header //
+if (document.getElementsByTagName('main')[0].classList.contains('home-page')) {
+  const header = document.getElementsByClassName('site-header')[0]
+  header.style.border = "3px solid white"
+}*/
+
+document.querySelectorAll('.btn-read-more').forEach(el => el.addEventListener('click', event => {
+  event.target.parentNode.querySelectorAll('p')[0].classList.toggle('expanded')
+  
+  if (event.target.parentNode.querySelectorAll('p')[0].classList.contains("expanded")) {
+    event.target.innerHTML = `Read less <span class="caret rotate-180"><svg xmlns="http://www.w3.org/2000/svg" width="16" viewBox="0 0 20 12" style="margin-left: 0.5rem;transform: rotate(180deg);"><path fill="#727272" d="m17.8.4-7.7 8.2L2.2.4C1.7-.1.9-.1.4.4s-.5 1.4 0 1.9l8.8 9.3c.3.3.7.4 1.1.4.3 0 .7-.1.9-.4l8.4-9.3c.5-.5.5-1.4 0-1.9s-1.3-.5-1.8 0z"/></svg></span>`;
+  } else {
+    event.target.innerHTML = `Read more <span class="caret"><svg xmlns="http://www.w3.org/2000/svg" width="16" viewBox="0 0 20 12" style="margin-left: 0.5rem;"><path fill="#727272" d="m17.8.4-7.7 8.2L2.2.4C1.7-.1.9-.1.4.4s-.5 1.4 0 1.9l8.8 9.3c.3.3.7.4 1.1.4.3 0 .7-.1.9-.4l8.4-9.3c.5-.5.5-1.4 0-1.9s-1.3-.5-1.8 0z"/></svg></span>`;
+  }
+})); 
+
+//site navigation
+/**
+ * Dropdown menu web component
+ *
+ * @element cagov-site-navigation
+ *
+ * @cssprop --primary-color - Default value of #064E66, used for background
+ * @cssprop --gray-300 - #e1e0e3
+ * @cssprop --primary-dark-color - #064e66
+ * @cssprop --secondary-color - #fec02f
+ * @cssprop --w-lg - '1176px'
+ */
+
+// Function determining if it's mobile view (max 767px)
+function mobileView() {
+  const mobileElement = document.querySelector(
+    '.site-header .grid-mobile-icons',
+  );
+  if (mobileElement) {
+    return getComputedStyle(mobileElement).display !== 'none';
+  }
+  return false;
+}
+class CAGovSiteNavigation extends window.HTMLElement {
+  connectedCallback() {
+    document
+      .querySelector('.cagov-nav.open-menu')
+      .addEventListener('click', this.toggleMainMenu.bind(this));
+
+    // Mobile search events
+    const mobileSearchBtn = document.querySelector(
+      '.cagov-nav.mobile-search .search-btn',
+    );
+    if (mobileSearchBtn) {
+      mobileSearchBtn.setAttribute('aria-expanded', 'false');
+      document
+        .querySelector('.search-container--small .site-search input')
+        .setAttribute('tabindex', '-1');
+      document
+        .querySelector(
+          '.search-container--small .site-search button.search-submit',
+        )
+        .setAttribute('tabindex', '-1');
+      document
+        .querySelector('.search-container--small')
+        .setAttribute('aria-hidden', 'true');
+      if (mobileView()) {
+        mobileSearchBtn.addEventListener('click', () => {
+          document
+            .querySelector('.search-container--small')
+            .classList.toggle('hidden-search');
+          const searchactive = document
+            .querySelector('.search-container--small')
+            .classList.contains('hidden-search');
+          if (searchactive) {
+            mobileSearchBtn.setAttribute('aria-expanded', 'false');
+            document
+              .querySelector('.search-container--small .site-search input')
+              .setAttribute('tabindex', '-1');
+            document
+              .querySelector(
+                '.search-container--small .site-search button.search-submit',
+              )
+              .setAttribute('tabindex', '-1');
+            document
+              .querySelector('.search-container--small')
+              .setAttribute('aria-hidden', 'true');
+          } else {
+            mobileSearchBtn.setAttribute('aria-expanded', 'true');
+            document
+              .querySelector('.search-container--small .site-search input')
+              .focus();
+            document
+              .querySelector('.search-container--small .site-search input')
+              .removeAttribute('tabindex');
+            document
+              .querySelector(
+                '.search-container--small .site-search button.search-submit',
+              )
+              .removeAttribute('tabindex');
+            document
+              .querySelector('.search-container--small')
+              .setAttribute('aria-hidden', 'false');
+          }
+        });
+      }
+    }
+
+    // reset mobile search on resize
+    window.addEventListener('resize', () => {
+      document
+        .querySelector('.search-container--small')
+        .classList.add('hidden-search');
+      if (mobileSearchBtn) {
+        document
+          .querySelector('.cagov-nav.mobile-search .search-btn')
+          .setAttribute('aria-expanded', 'false');
+      }
+      document
+        .querySelector('.search-container--small .site-search input')
+        .setAttribute('tabindex', '-1');
+      document
+        .querySelector(
+          '.search-container--small .site-search button.search-submit',
+        )
+        .setAttribute('tabindex', '-1');
+      document
+        .querySelector('.search-container--small')
+        .setAttribute('aria-hidden', 'true');
+      // reset navigation on resize
+      this.closeAllMenus();
+      this.closeMainMenu();
+    });
+
+    this.expansionListeners();
+    document.addEventListener('keydown', this.escapeMainMenu.bind(this));
+    document.body.addEventListener('click', this.bodyClick.bind(this));
+    this.highlightCurrentPage();
+  }
+
+  toggleMainMenu() {
+    if (
+      document
+        .querySelector('.cagov-nav.hamburger')
+        .classList.contains('is-active')
+    ) {
+      this.closeMainMenu();
+    } else {
+      this.openMainMenu();
+    }
+  }
+
+  highlightCurrentPage() {
+    this.querySelectorAll('a.expanded-menu-dropdown-link').forEach((link) => {
+      if (link.href === window.location.href) {
+        link.classList.add('current-page-highlight');
+      }
+    });
+  }
+
+  openMainMenu() {
+    document.querySelector('.mobile-icons').classList.add('display-menu');
+    this.classList.add('display-menu');
+    document.querySelector('.cagov-nav.hamburger').classList.add('is-active');
+    document.querySelector('.cagov-nav.menu-trigger').classList.add('is-fixed');
+    document
+      .querySelector('.cagov-nav.menu-trigger')
+      .setAttribute('aria-expanded', 'true');
+    const menLabel = document.querySelector('.cagov-nav.menu-trigger-label');
+    menLabel.innerHTML = menLabel.getAttribute('data-closelabel');
+  }
+
+  closeMainMenu() {
+    document.querySelector('.mobile-icons').classList.remove('display-menu');
+    this.classList.remove('display-menu');
+    document
+      .querySelector('.cagov-nav.hamburger')
+      .classList.remove('is-active');
+    document
+      .querySelector('.cagov-nav.menu-trigger')
+      .classList.remove('is-fixed');
+    document
+      .querySelector('.cagov-nav.menu-trigger')
+      .setAttribute('aria-expanded', 'false');
+    const menLabel = document.querySelector('.cagov-nav.menu-trigger-label');
+    menLabel.innerHTML = menLabel.getAttribute('data-openlabel');
+  }
+
+  escapeMainMenu(event) {
+    // Close menus if user presses escape key.
+    if (event.keyCode === 27) {
+      this.closeAllMenus();
+    }
+  }
+
+  bodyClick(event) {
+    if (!event.target.closest('cagov-site-navigation')) {
+      this.closeAllMenus();
+    }
+  }
+
+  closeAllMenus() {
+    const allMenus = this.querySelectorAll('.js-cagov-navoverlay-expandable');
+    allMenus.forEach((menu) => {
+      const expandedEl = menu.querySelector('.expanded-menu-section');
+      expandedEl.classList.remove('expanded');
+      menu.setAttribute('aria-expanded', 'false');
+      const closestDropDown = menu.querySelector('.expanded-menu-dropdown');
+      if (closestDropDown) {
+        closestDropDown.setAttribute('aria-hidden', 'true');
+        const allLinks = closestDropDown.querySelectorAll('a');
+        allLinks.forEach((link) => {
+          link.setAttribute('tabindex', '-1'); // set tabindex to -1 so you cannot tab through these hidden links
+        });
+      }
+    });
+  }
+
+  expansionListeners() {
+    const allMenus = this.querySelectorAll('.js-cagov-navoverlay-expandable');
+    allMenus.forEach((menu) => {
+      const nearestMenu = menu.querySelector('.expanded-menu-section');
+      if (nearestMenu) {
+        const nearestMenuDropDown = nearestMenu.querySelector(
+          '.expanded-menu-dropdown',
+        );
+        if (nearestMenuDropDown) {
+          nearestMenuDropDown.setAttribute('aria-hidden', 'true');
+          menu.setAttribute('aria-expanded', 'false');
+        }
+      }
+      const menuComponent = this;
+      menu.addEventListener('click', function addingClickListener(event) {
+        if (event.target.nodeName !== 'A') {
+          event.preventDefault();
+        }
+        const expandedEl = this.querySelector('.expanded-menu-section');
+        if (expandedEl) {
+          if (expandedEl.classList.contains('expanded')) {
+            // closing an open menu
+            menuComponent.closeAllMenus();
+          } else {
+            menuComponent.closeAllMenus();
+            expandedEl.classList.add('expanded');
+            menu.setAttribute('aria-expanded', 'true');
+            const closestDropDown = this.querySelector(
+              '.expanded-menu-dropdown',
+            );
+            if (closestDropDown) {
+              closestDropDown.setAttribute('aria-hidden', 'false');
+              const allLinks = closestDropDown.querySelectorAll('a');
+              allLinks.forEach((link) => {
+                link.removeAttribute('tabindex'); // remove tabindex from all the links
+              });
+            }
+          }
+        }
+      });
+    });
+  }
+}
+window.customElements.define('cagov-site-navigation', CAGovSiteNavigation);
+
+//Page Navigation
 var styles =
   '/* PAGE NAVIGATION */\nsidebar cagov-page-navigation .label {\n  font-weight: 700;\n  font-size: 24px;\n  line-height: 28.2px;\n  padding: 0;\n  margin: 0;\n  padding-bottom: 16px;\n}\n\nsidebar cagov-page-navigation ul,\nsidebar cagov-page-navigation ol:not([class*=menu]):not([class*=nav]):not([class*=footer-links]),\nsidebar cagov-page-navigation ul:not([class*=menu]):not([class*=nav]):not([class*=footer-links]) {\n  margin: 0;\n  text-indent: 0;\n  padding: 0;\n}\n\nsidebar cagov-page-navigation ul li {\n  padding-top: 14px;\n  padding-bottom: 18px;\n  margin-left: 0;\n  margin-top: 0px;\n  margin-bottom: 0px;\n  border-bottom: 1px solid var(--gray-300, #e1e0e3);\n  line-height: 28.2px;\n  list-style: none;\n}\nsidebar cagov-page-navigation ul li:first-child {\n  border-top: 1px solid var(--gray-300, #e1e0e3);\n}\nsidebar cagov-page-navigation ul li a {\n  text-decoration: none;\n}\nsidebar cagov-page-navigation ul li a:hover {\n  text-decoration: underline;\n}\n\n@media only screen and (max-width: 992px) {\n  cagov-page-navigation .label {\n    display: none;\n  }\n\n  .sidebar-container {\n    display: block;\n    width: 100%;\n    max-width: 100%;\n  }\n\n  cagov-page-navigation ul li a {\n    font-size: 16px;\n    line-height: 24px;\n  }\n}\n\n/*# sourceMappingURL=index.css.map */\n';
